data-layer/projects/proj_004/phase2.4_implementation/rag_system/requirements.txt
# RAG System MVP - Phase 2.4
# Python 3.10+

# 核心依赖
openai>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
pyyaml>=6.0

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# 工具
tqdm>=4.66.0
python-dotenv>=1.0.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
